Move BNB token query and transfer out of unit test
