+++ tools/anda_bnb_examples/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "anda_bnb_examples"
+version = "0.1.0"
+description.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anda_bnb = { path = "../anda_bnb" }
+anda_core = { path = "../../anda_core" }
+anda_engine = { path = "../../anda_engine" }
+anda_web3_client = { path = "../../anda_web3_client" }
+anda_engine_server = { path = "../../anda_engine_server" }
+serde = { workspace = true }
+serde_json = { workspace = true }
+schemars = { workspace = true }
+tokio = { workspace = true }
+log = { workspace = true }
+dotenv = { workspace = true }
+reqwest = { workspace = true }
+alloy = { version = "0.14.0", features = [
+    "providers", "network", "sol-types", "consensus", "signers",
+] }
+thiserror = "1.0"
+k256 = { version = "0.13", features = ["ecdsa"] }
+async-trait = "0.1"
+structured-logger = { workspace = true }
+rand = { workspace = true }
+
+[dev-dependencies]
